+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-
+## [develop] - 22/02/2023
+- add table payment_cards
+- composer update
+- 
 ## [develop,test,master] - 31/01/2022
 - php artisan update:db